Whatever they were doing to retrieve blogs was bad, I've changed it all, get_blogs.php now handles all getting of blogs in any way
